✨ feat(fish): missing_package function to use when calling an app thats not installed in the system

+++ .config/fish/functions/fe.fish
@@ -1,8 +1,12 @@
 function fe
-    set tmp (mktemp -t "yazi-cwd.XXXXX")
-    yazi --cwd-file="$tmp"
-    if set cwd (cat -- "$tmp"); and [ -n "$cwd" ]; and [ "$cwd" != "$PWD" ]
-        cd -- "$cwd"
+    if type -f yazi &>/dev/null
+        set tmp (mktemp -t "yazi-cwd.XXXXX")
+        yazi --cwd-file="$tmp"
+        if set cwd (cat -- "$tmp"); and [ -n "$cwd" ]; and [ "$cwd" != "$PWD" ]
+            cd -- "$cwd"
+        end
+        rm -f -- "$tmp"
+    else
+        missing_package yazi
     end
-    rm -f -- "$tmp"
 end
