+++ .config/fish/functions/ni.fish
@@ -1,3 +1,7 @@
 function ni --wraps='npm install' --description 'alias ni=npm install'
-    npm install $argv
+    if type -f npm &>/dev/null
+        npm install $argv
+    else
+        missing_package npm
+    end
 end
