+++ .config/fish/functions/nid.fish
@@ -1,3 +1,7 @@
 function nid --wraps='npm install --save-dev' --description 'alias nid=npm install --save-dev'
-    npm install --save-dev $argv
+    if type -f npm &>/dev/null
+        npm install --save-dev $argv
+    else
+        missing_package npm
+    end
 end
