+++ .config/fish/functions/nis.fish
@@ -1,3 +1,7 @@
 function nis --wraps='npm install --save' --description 'alias nis=npm install --save'
-    npm install --save $argv
+    if type -f npm &>/dev/null
+        npm install --save $argv
+    else
+        missing_package npm
+    end
 end
