+++ .config/fish/functions/nu.fish
@@ -1,3 +1,7 @@
 function nu --wraps='npm uninstall' --description 'alias nu=npm uninstall'
-    npm uninstall $argv
+    if type -f npm &>/dev/null
+        npm uninstall $argv
+    else
+        missing_package npm
+    end
 end
