+++ .config/fish/functions/nud.fish
@@ -1,3 +1,7 @@
 function nud --wraps='npm uninstall --save-dev' --description 'alias nud=npm uninstall --save-dev'
-    npm uninstall --save-dev $argv
+    if type -f npm &>/dev/null
+        npm uninstall --save-dev $argv
+    else
+        missing_package npm
+    end
 end
