+++ .config/fish/functions/nus.fish
@@ -1,3 +1,7 @@
 function nus --wraps='npm uninstall --save' --description 'alias nus=npm uninstall --save'
-    npm uninstall --save $argv
+    if type -f npm &>/dev/null
+        npm uninstall --save $argv
+    else
+        missing_package npm
+    end
 end
